refactor(services): move some services into subfolders
