+++ src/features/discord/messages/services/command/feature/features/noon/classes/discord-message-command-feature-noon-help.ts
@@ -8,7 +8,7 @@ import { discordGetCommandAndFirstArgument } from '../../../../../../functions/c
 import { IDiscordMessageResponse } from '../../../../../../interfaces/discord-message-response';
 import { IAnyDiscordMessage } from '../../../../../../types/any-discord-message';
 import { DiscordMessageConfigService } from '../../../../../config/discord-message-config.service';
-import { DiscordMessageHelpService } from '../../../../../discord-message-help.service';
+import { DiscordMessageHelpService } from '../../../../../helpers/discord-message-help.service';
 import { EmbedFieldData, MessageEmbedOptions, Snowflake } from 'discord.js';
 import _ from 'lodash';
