+++ src/features/discord/messages/services/helpers/discord-message-content.service.ts
@@ -1,5 +1,5 @@
-import { AbstractService } from '../../../../classes/services/abstract.service';
-import { ServiceNameEnum } from '../../../../enums/service-name.enum';
+import { AbstractService } from '../../../../../classes/services/abstract.service';
+import { ServiceNameEnum } from '../../../../../enums/service-name.enum';
 import _ from 'lodash';
 
 export class DiscordMessageContentService extends AbstractService {
