+++ src/features/discord/messages/services/helpers/discord-message-error.service.spec.ts
@@ -1,6 +1,6 @@
 import { DiscordMessageErrorService } from './discord-message-error.service';
-import { ServiceNameEnum } from '../../../../enums/service-name.enum';
-import { CoreEventService } from '../../../core/services/core-event.service';
+import { ServiceNameEnum } from '../../../../../enums/service-name.enum';
+import { CoreEventService } from '../../../../core/services/core-event.service';
 
 describe(`DiscordMessageErrorService`, (): void => {
   let service: DiscordMessageErrorService;
