+++ src/features/discord/messages/services/responses/discord-message-author.service.spec.ts
@@ -1,6 +1,6 @@
 import { DiscordMessageAuthorService } from './discord-message-author.service';
-import { ServiceNameEnum } from '../../../../enums/service-name.enum';
-import { CoreEventService } from '../../../core/services/core-event.service';
+import { ServiceNameEnum } from '../../../../../enums/service-name.enum';
+import { CoreEventService } from '../../../../core/services/core-event.service';
 
 describe(`DiscordMessageAuthorService`, (): void => {
   let service: DiscordMessageAuthorService;
