+++ src/features/discord/messages/services/types/discord-message-text.service.spec.ts
@@ -1,6 +1,6 @@
 import { DiscordMessageTextService } from './discord-message-text.service';
-import { ServiceNameEnum } from '../../../../enums/service-name.enum';
-import { CoreEventService } from '../../../core/services/core-event.service';
+import { ServiceNameEnum } from '../../../../../enums/service-name.enum';
+import { CoreEventService } from '../../../../core/services/core-event.service';
 
 describe(`DiscordMessageTextService`, (): void => {
   let service: DiscordMessageTextService;
