+++ src/features/discord/messages/services/types/discord-message-text.service.ts
@@ -1,20 +1,20 @@
-import { DiscordMessageCommandService } from './command/discord-message-command.service';
-import { DiscordMessageAuthorService } from './discord-message-author.service';
-import { DiscordMessageContentService } from './discord-message-content.service';
-import { AbstractService } from '../../../../classes/services/abstract.service';
-import { ServiceNameEnum } from '../../../../enums/service-name.enum';
-import { AppConfigService } from '../../../app/services/config/app-config.service';
-import { LoggerService } from '../../../logger/services/logger.service';
-import { ProfileConfigService } from '../../../profile/services/config/profile-config.service';
-import { addDiscordDevPrefix } from '../../functions/dev-prefix/add-discord-dev-prefix';
-import { DiscordMentionService } from '../../mentions/services/discord-mention.service';
-import { DiscordAuthorService } from '../../users/services/discord-author.service';
-import { DiscordSoniaService } from '../../users/services/discord-sonia.service';
-import { ISonia } from '../../users/types/sonia';
-import { isDiscordMessage } from '../functions/is-discord-message';
-import { IDiscordMessageResponse } from '../interfaces/discord-message-response';
-import { IAnyDiscordMessage } from '../types/any-discord-message';
-import { IDiscordMessage } from '../types/discord-message';
+import { AbstractService } from '../../../../../classes/services/abstract.service';
+import { ServiceNameEnum } from '../../../../../enums/service-name.enum';
+import { AppConfigService } from '../../../../app/services/config/app-config.service';
+import { LoggerService } from '../../../../logger/services/logger.service';
+import { ProfileConfigService } from '../../../../profile/services/config/profile-config.service';
+import { addDiscordDevPrefix } from '../../../functions/dev-prefix/add-discord-dev-prefix';
+import { DiscordMentionService } from '../../../mentions/services/discord-mention.service';
+import { DiscordAuthorService } from '../../../users/services/discord-author.service';
+import { DiscordSoniaService } from '../../../users/services/discord-sonia.service';
+import { ISonia } from '../../../users/types/sonia';
+import { isDiscordMessage } from '../../functions/is-discord-message';
+import { IDiscordMessageResponse } from '../../interfaces/discord-message-response';
+import { IAnyDiscordMessage } from '../../types/any-discord-message';
+import { IDiscordMessage } from '../../types/discord-message';
+import { DiscordMessageCommandService } from '../command/discord-message-command.service';
+import { DiscordMessageContentService } from '../helpers/discord-message-content.service';
+import { DiscordMessageAuthorService } from '../responses/discord-message-author.service';
 import _ from 'lodash';
 
 export class DiscordMessageTextService extends AbstractService {
